+++ source/modifiers/quicken.js
@@ -2,7 +2,7 @@ const { ModifierTemplate } = require("../classes");
 
 module.exports = new ModifierTemplate("Quicken",
 	"Increase move speed by @{stackCount*5} for @{stackCount} rounds.",
-	false,
 	true,
+	false,
 	0
 ).setInverse("Slow");